+++ shimmer/__init__.py
@@ -31,6 +31,18 @@ from shimmer.modules.losses import (
     GWLossesWithUncertainty,
     LossCoefs,
 )
+from shimmer.modules.utils import (
+    batch_cycles,
+    batch_cycles_with_uncertainty,
+    batch_demi_cycles,
+    batch_demi_cycles_with_uncertainty,
+    batch_translations,
+    batch_translations_with_uncertainty,
+    cycle,
+    cycle_with_uncertainty,
+    translation,
+    translation_with_uncertainty,
+)
 from shimmer.types import (
     LatentsDomainGroupDT,
     LatentsDomainGroupsDT,
@@ -81,4 +93,14 @@ __all__ = [
     "GWLosses",
     "GWLossesWithUncertainty",
     "RepeatedDataset",
+    "batch_cycles",
+    "batch_demi_cycles",
+    "batch_translations",
+    "cycle",
+    "translation",
+    "cycle_with_uncertainty",
+    "translation_with_uncertainty",
+    "batch_translations_with_uncertainty",
+    "batch_demi_cycles_with_uncertainty",
+    "batch_cycles_with_uncertainty",
 ]