+++ shimmer/modules/__init__.py
@@ -31,6 +31,18 @@ from shimmer.modules.losses import (
     GWLossesWithUncertainty,
     LossCoefs,
 )
+from shimmer.modules.utils import (
+    batch_cycles,
+    batch_cycles_with_uncertainty,
+    batch_demi_cycles,
+    batch_demi_cycles_with_uncertainty,
+    batch_translations,
+    batch_translations_with_uncertainty,
+    cycle,
+    cycle_with_uncertainty,
+    translation,
+    translation_with_uncertainty,
+)
 from shimmer.modules.vae import (
     VAE,
     VAEDecoder,
@@ -73,4 +85,14 @@ __all__ = [
     "VAEEncoder",
     "VAEDecoder",
     "VAE",
+    "batch_cycles",
+    "batch_demi_cycles",
+    "batch_translations",
+    "cycle",
+    "translation",
+    "cycle_with_uncertainty",
+    "translation_with_uncertainty",
+    "batch_translations_with_uncertainty",
+    "batch_demi_cycles_with_uncertainty",
+    "batch_cycles_with_uncertainty",
 ]