+++ shimmer/modules/global_workspace.py
@@ -28,6 +28,14 @@ from shimmer.modules.losses import (
     GWLossesWithUncertainty,
     LossCoefs,
 )
+from shimmer.modules.utils import (
+    batch_cycles,
+    batch_cycles_with_uncertainty,
+    batch_demi_cycles,
+    batch_demi_cycles_with_uncertainty,
+    batch_translations,
+    batch_translations_with_uncertainty,
+)
 from shimmer.types import (
     LatentsDomainGroupsDT,
     LatentsDomainGroupsT,
@@ -37,6 +45,7 @@ from shimmer.types import (
     RawDomainGroupsT,
     RawDomainGroupT,
 )
+from shimmer.utils import groups_batch_size
 
 
 class SchedulerArgs(TypedDict, total=False):
@@ -49,26 +58,8 @@ class SchedulerArgs(TypedDict, total=False):
     """Total number of steps"""
 
 
-class GWPredictions(TypedDict):
+class GWPredictionsBase(TypedDict):
     """TypedDict of the output given when calling `GlobalWorkspaceBase.predict`"""
-
-    demi_cycles: dict[str, torch.Tensor]
-    """Demi-cycle predictions of the model for each domain. Only computed on domain
-    groups with only one domain.
-    """
-
-    cycles: dict[tuple[str, str], torch.Tensor]
-    """Cycle predictions of the model from one domain through another one.
-    Only computed on domain groups with more than one domain.
-    The keys are tuple with start domain and intermediary domain.
-    """
-
-    translations: dict[tuple[str, str], torch.Tensor]
-    """Translation predictions of the model from one domain through another one.
-
-    Only computed on domain groups with more than one domain.
-    The keys are tuples with start domain and target domain.
-    """
 
     states: dict[str, torch.Tensor]
     """GW state representation from domain groups with only one domain.
@@ -173,22 +164,17 @@ class GlobalWorkspaceBase(LightningModule):
     def forward(  # type: ignore
         self,
         latent_domains: LatentsDomainGroupsT,
-    ) -> GWPredictions:
+    ) -> GWPredictionsBase:
         """Computes demi-cycles, cycles, and translations.
 
         Args:
             latent_domains (`LatentsT`): Groups of domains for the computation.
 
         Returns:
-            `GWPredictions`: the predictions on the batch.
+            `GWPredictionsBase`: the predictions on the batch.
         """
 
-        return GWPredictions(
-            demi_cycles=self.batch_demi_cycles(latent_domains),
-            cycles=self.batch_cycles(latent_domains),
-            translations=self.batch_translations(latent_domains),
-            states=self.batch_gw_states(latent_domains),
-        )
+        return GWPredictionsBase(states=self.batch_gw_states(latent_domains))
 
     def batch_gw_states(
         self, latent_domains: LatentsDomainGroupsT
@@ -208,78 +194,6 @@ class GlobalWorkspaceBase(LightningModule):
             domain_name = list(domains)[0]
             z = self.gw_mod.encode(latents)
             predictions[domain_name] = z
-        return predictions
-
-    def batch_demi_cycles(
-        self, latent_domains: LatentsDomainGroupsT
-    ) -> dict[str, torch.Tensor]:
-        """Computes demi-cycles of a batch of groups of domains.
-
-        Args:
-            latent_domains (`LatentsT`): the batch of groups of domains
-
-        Returns:
-            `dict[str, torch.Tensor]`: demi-cycles predictions for each domain.
-        """
-        predictions: dict[str, torch.Tensor] = {}
-        for domains, latents in latent_domains.items():
-            if len(domains) > 1:
-                continue
-            domain_name = list(domains)[0]
-            z = self.gw_mod.translate(latents, to=domain_name)
-            predictions[domain_name] = z
-        return predictions
-
-    def batch_cycles(
-        self, latent_domains: LatentsDomainGroupsT
-    ) -> dict[tuple[str, str], torch.Tensor]:
-        """Computes cycles of a batch of groups of domains.
-
-        Args:
-            latent_domains (`LatentsT`): the batch of groups of domains
-
-        Returns:
-            `dict[tuple[str, str], torch.Tensor]`: cycles predictions for each
-                couple of (start domain, intermediary domain).
-        """
-        predictions: dict[tuple[str, str], torch.Tensor] = {}
-        for domains_source, latents_source in latent_domains.items():
-            if len(domains_source) > 1:
-                continue
-            domain_name_source = next(iter(domains_source))
-            for domain_name_target in self.domain_mods.keys():
-                if domain_name_source == domain_name_target:
-                    continue
-                z = self.gw_mod.cycle(latents_source, through=domain_name_target)
-                domains = (domain_name_source, domain_name_target)
-                predictions[domains] = z[domain_name_source]
-        return predictions
-
-    def batch_translations(
-        self, latent_domains: LatentsDomainGroupsT
-    ) -> dict[tuple[str, str], torch.Tensor]:
-        """Computes translations of a batch of groups of domains.
-
-        Args:
-            latent_domains (`LatentsT`): the batch of groups of domains
-
-        Returns:
-            `dict[tuple[str, str], torch.Tensor]`: translation predictions for each
-                couple of (start domain, target domain).
-        """
-        predictions: dict[tuple[str, str], torch.Tensor] = {}
-        for domains, latents in latent_domains.items():
-            if len(domains) < 2:
-                continue
-            for domain_name_source in domains:
-                for domain_name_target in domains:
-                    if domain_name_source == domain_name_target:
-                        continue
-                    prediction = self.gw_mod.translate(
-                        {domain_name_source: latents[domain_name_source]},
-                        to=domain_name_target,
-                    )
-                    predictions[(domain_name_source, domain_name_target)] = prediction
         return predictions
 
     def encode_domain(self, domain: Any, name: str) -> torch.Tensor:
@@ -356,20 +270,6 @@ class GlobalWorkspaceBase(LightningModule):
             for domains, latents in latents_domain.items()
         }
 
-    def _get_batch_size(self, domain_latents: LatentsDomainGroupsT) -> int:
-        """Get the batch size of the batch.
-
-        Args:
-            domain_latents (`LatentsDomainGroupsT`): the batch of groups.
-
-        Returns:
-            int: the batch size.
-        """
-        for data in domain_latents.values():
-            for tensor in data.values():
-                return tensor.size(0)
-        raise ValueError("Empty batch.")
-
     def generic_step(self, batch: RawDomainGroupsT, mode: ModelModeT) -> torch.Tensor:
         """The generic step used in `training_step`, `validation_step` and
         `test_step`.
@@ -382,7 +282,7 @@ class GlobalWorkspaceBase(LightningModule):
             `torch.Tensor`: the loss to train on.
         """
         domain_latents = self.encode_domains(batch)
-        batch_size = self._get_batch_size(domain_latents)
+        batch_size = groups_batch_size(domain_latents)
 
         loss_output = self.loss_mod.step(domain_latents, mode)
 
@@ -429,7 +329,7 @@ class GlobalWorkspaceBase(LightningModule):
 
     def predict_step(  # type: ignore
         self, data: Mapping[str, Any], batch_idx: int
-    ) -> GWPredictions:
+    ) -> GWPredictionsBase:
         """Predict step used by lightning"""
 
         batch = {frozenset(data.keys()): data}
@@ -483,6 +383,28 @@ def freeze_domain_modules(
         mod.freeze()
     # Cast for better auto-completion at the expense of ModuleDict
     return cast(dict[str, DomainModule], ModuleDict(domain_mods))
+
+
+class GWPredictions(GWPredictionsBase):
+    """TypedDict of the output given when calling `GlobalWorkspaceBase.predict`"""
+
+    demi_cycles: dict[str, torch.Tensor]
+    """Demi-cycle predictions of the model for each domain. Only computed on domain
+    groups with only one domain.
+    """
+
+    cycles: dict[tuple[str, str], torch.Tensor]
+    """Cycle predictions of the model from one domain through another one.
+    Only computed on domain groups with more than one domain.
+    The keys are tuple with start domain and intermediary domain.
+    """
+
+    translations: dict[tuple[str, str], torch.Tensor]
+    """Translation predictions of the model from one domain through another one.
+
+    Only computed on domain groups with more than one domain.
+    The keys are tuples with start domain and target domain.
+    """
 
 
 class GlobalWorkspace(GlobalWorkspaceBase):
@@ -539,6 +461,25 @@ class GlobalWorkspace(GlobalWorkspaceBase):
 
         super().__init__(gw_mod, loss_mod, optim_lr, optim_weight_decay, scheduler_args)
 
+    def forward(  # type: ignore
+        self,
+        latent_domains: LatentsDomainGroupsT,
+    ) -> GWPredictions:
+        """Computes demi-cycles, cycles, and translations.
+
+        Args:
+            latent_domains (`LatentsT`): Groups of domains for the computation.
+
+        Returns:
+            `GWPredictions`: the predictions on the batch.
+        """
+        return GWPredictions(
+            demi_cycles=batch_demi_cycles(self.gw_mod, latent_domains),
+            cycles=batch_cycles(self.gw_mod, latent_domains, self.domain_mods.keys()),
+            translations=batch_translations(self.gw_mod, latent_domains),
+            **super().forward(latent_domains),
+        )
+
 
 class GlobalWorkspaceWithUncertainty(GlobalWorkspaceBase):
     """A simple 2-domains max GlobalWorkspaceBase with uncertainty.
@@ -546,6 +487,8 @@ class GlobalWorkspaceWithUncertainty(GlobalWorkspaceBase):
     This is used to simplify a Global Workspace instanciation and only overrides the
     `__init__` method.
     """
+
+    gw_mod: GWModuleWithUncertainty
 
     def __init__(
         self,
@@ -619,6 +562,29 @@ class GlobalWorkspaceWithUncertainty(GlobalWorkspaceBase):
         )
 
         super().__init__(gw_mod, loss_mod, optim_lr, optim_weight_decay, scheduler_args)
+
+    def forward(  # type: ignore
+        self,
+        latent_domains: LatentsDomainGroupsT,
+    ) -> GWPredictions:
+        """Computes demi-cycles, cycles, and translations.
+
+        Args:
+            latent_domains (`LatentsT`): Groups of domains for the computation.
+
+        Returns:
+            `GWPredictions`: the predictions on the batch.
+        """
+        return GWPredictions(
+            demi_cycles=batch_demi_cycles_with_uncertainty(self.gw_mod, latent_domains),
+            cycles=batch_cycles_with_uncertainty(
+                self.gw_mod, latent_domains, self.domain_mods.keys()
+            ),
+            translations=batch_translations_with_uncertainty(
+                self.gw_mod, latent_domains
+            ),
+            **super().forward(latent_domains),
+        )
 
 
 class GlobalWorkspaceFusion(GlobalWorkspaceBase):